新需求开发，数据同步完善

+++ app/src/main/java/com/cec/zbgl/service/SyncService.java
@@ -2,15 +2,18 @@ package com.cec.zbgl.service;
 
 import android.app.Activity;
 import android.content.Intent;
+import android.support.constraint.solver.LinearSystem;
 
 import com.cec.zbgl.dto.CourseDto;
 import com.cec.zbgl.dto.DeviceDto;
+import com.cec.zbgl.dto.DeviceReleDto;
 import com.cec.zbgl.dto.OrgnizationDto;
 import com.cec.zbgl.dto.SyncDto;
 import com.cec.zbgl.dto.UserDto;
 import com.cec.zbgl.event.MessageEvent;
 import com.cec.zbgl.model.DeviceCourse;
 import com.cec.zbgl.model.DeviceInfo;
+import com.cec.zbgl.model.DeviceRele;
 import com.cec.zbgl.model.ServerConfig;
 import com.cec.zbgl.model.SpOrgnization;
 import com.cec.zbgl.model.User;
@@ -33,6 +36,8 @@ import java.io.PrintStream;
 import java.net.Socket;
 import java.util.ArrayList;
 import java.util.List;
+import java.util.concurrent.ExecutorService;
+import java.util.concurrent.Executors;
 
 import okhttp3.Call;
 import okhttp3.Callback;
@@ -55,12 +60,14 @@ public class SyncService {
     private List<DeviceDto> dList;
     private List<CourseDto> cList;
     private List<OrgnizationDto> oList;
+    private List<DeviceReleDto> rList;
     private List<UserDto> uList;
     private SyncDto syncDto;
     private Gson gson;
     private Activity mActivity;
     private String adress;
     private int port;
+    private List<String> mList;
 
 
 
@@ -72,9 +79,14 @@ public class SyncService {
         gson = new Gson();
         syncDto = new SyncDto();
         mActivity = activity;
+        mList = new ArrayList<>();
     }
 
 
+    /**
+     * 通过RESTFUL进行数据同步
+     * 暂弃用
+     */
     public void syncData() {
         MediaType mediaType = MediaType.parse("application/json; charset=utf-8");
         ServerService service = new ServerService();
@@ -127,8 +139,6 @@ public class SyncService {
 
                 response.body().close();
                 EventBus.getDefault().post(new MessageEvent("刷新UI"));
-//                Intent intent = new Intent("MY_BROADCAST");
-//                mActivity.sendBroadcast(intent);
             }
         });
     }
@@ -156,9 +166,12 @@ public class SyncService {
     }
 
 
-
-    public void socketSync(){
+    /**
+     * 通过socket进行数据同步
+     */
+    public void socketSync(List<String> mList){
         service = new ServerService();
+        this.mList = mList;
         ServerConfig config = service.load();
         OutputStream os = null;
         InputStream is = null;
@@ -167,14 +180,14 @@ public class SyncService {
             adress = config.getIp();
             port = 8088;
         }else {
-            EventBus.getDefault().post(new MessageEvent("刷新UI"));
+            EventBus.getDefault().post(new MessageEvent("failed"));
             return;
         }
 
         try {
             socket = new Socket(adress, 8088);
             if (!socket.isConnected()) {
-                EventBus.getDefault().post(new MessageEvent("刷新UI"));
+                EventBus.getDefault().post(new MessageEvent("failed"));
                 return;
             }
             os = socket.getOutputStream();
@@ -189,8 +202,10 @@ public class SyncService {
             reader.close();
             is.close();
             out.close();
+            EventBus.getDefault().post(new MessageEvent("succeed"));
         } catch (IOException e) {
             e.printStackTrace();
+            EventBus.getDefault().post(new MessageEvent("failed"));
         } finally {
             if (socket != null) {
                 try {
@@ -199,7 +214,7 @@ public class SyncService {
                     e.printStackTrace();
                 }
             }
-            EventBus.getDefault().post(new MessageEvent("刷新UI"));
+
         }
     }
 
@@ -209,38 +224,82 @@ public class SyncService {
      * @param out
      */
     public void sendMsg(PrintStream out) {
-        int flag = 0; int flag0 = 0; int count = 0; int count0 = 0;
-        //备品信息
-        count0 = deviceService.getCount();
-        while (flag0 <= count0) {
+
+        if (mList.size() > 0) {
             SyncDto syncDto = new SyncDto();
-            List<DeviceDto> dDtoList = new ArrayList<>();
-            List<DeviceInfo> deviceInfos = deviceService.loadByPage(flag0);
-            deviceInfos.stream().forEach( deviceInfo -> {
-                DeviceDto dto = DtoUtils.toDeviceDto(deviceInfo);
-                dDtoList.add(dto);
-            });
-            syncDto.setdList(dDtoList);
+            syncDto.setmList(mList);
             out.println(gson.toJson(syncDto));
             out.flush();
-            flag0 += 20;
+        }
+        int flag = 0; int flag0 = 0; int count = 0; int count0 = 0;
+
+        //发送设备信息
+        if (mList.contains("device")) {
+            //待同步备品信息总数
+            count0 = deviceService.getCount();
+            while (flag0 <= count0) {
+                SyncDto syncDto = new SyncDto();
+                List<DeviceDto> dDtoList = new ArrayList<>();
+                List<DeviceInfo> deviceInfos = deviceService.loadByPage(flag0);
+                deviceInfos.stream().forEach( deviceInfo -> {
+                    DeviceDto dto = DtoUtils.toDeviceDto(deviceInfo);
+                    dDtoList.add(dto);
+                });
+                syncDto.setdList(dDtoList);
+                out.println(gson.toJson(syncDto));
+                out.flush();
+                flag0 += 20;
+            }
+
+            //设备关联信息
+            int count1 = deviceService.getCountbyRele();
+            List<DeviceReleDto> rDtoList = new ArrayList<>();
+            List<DeviceRele> releList = deviceService.getSyncReleList();
+            releList.stream().forEach(rele -> {
+                DeviceReleDto releDto = DtoUtils.toReleDto(rele);
+                rDtoList.add(releDto);
+            });
+            syncDto.setrList(rDtoList);
+            out.println(gson.toJson(syncDto));
+            out.flush();
         }
 
-        //教程信息
-        count = courseService.getCount();
-        flag = 0;
-        while (flag <= count) {
-            SyncDto syncDto = new SyncDto();
-            List<CourseDto> cDtoList = new ArrayList<>();
-            List<DeviceCourse> courses = courseService.loadByPage(flag);
-            courses.stream().forEach( course -> {
-                CourseDto dto = DtoUtils.toCourseDto(course);
-                cDtoList.add(dto);
-            });
-            syncDto.setcList(cDtoList);
-            out.println(gson.toJson(syncDto));
-            out.flush();
-            flag += 20;
+        //发送教程信息
+        if (mList.contains("course")) {
+            //待同步教程信息总数
+            count = courseService.getCount();
+            flag = 0;
+            while (flag <= count) {
+                SyncDto syncDto = new SyncDto();
+                List<CourseDto> cDtoList = new ArrayList<>();
+                List<DeviceCourse> courses = courseService.loadByPage(flag);
+                courses.stream().forEach( course -> {
+                    CourseDto dto = DtoUtils.toCourseDto(course);
+                    cDtoList.add(dto);
+                });
+                syncDto.setcList(cDtoList);
+                out.println(gson.toJson(syncDto));
+                out.flush();
+                flag += 20;
+            }
+        }
+
+        if (mList.contains("orgnization")) {
+            int count2 = orgsService.getCount();
+            int flag2 = 0;
+            while (flag2 <= count2) {
+                SyncDto syncDto = new SyncDto();
+                List<OrgnizationDto> oDtoList = new ArrayList<>();
+                List<SpOrgnization> oList = orgsService.loadByPage(flag2);
+                oList.stream().forEach( org -> {
+                    OrgnizationDto dto = DtoUtils.toOrgDto(org);
+                    oDtoList.add(dto);
+                });
+                syncDto.setoList(oDtoList);
+                out.println(gson.toJson(syncDto));
+                out.flush();
+                flag2 += 20;
+            }
         }
     }
 
@@ -252,11 +311,23 @@ public class SyncService {
      */
     public void receiveMsg(BufferedReader br) throws IOException {
         String result = "";
+        if (mList.contains("orgnization")) {
+            orgsService.deleteAll();
+        }
 
-        //清空旧数据
-        courseService.deleteAll();
-        deviceService.deleteAll();
-        orgsService.deleteAll();
+        if (mList.contains("device")) {
+            deviceService.deleteAll();
+            deviceService.deleteReleAll();
+        }
+
+        if (mList.contains("course")) {
+            courseService.deleteAll();
+        }
+
+        if (mList.contains("user")) {
+            //TODO 删除人员信息
+        }
+
         //插入服务端同步数据
         while ( (result = br.readLine()) != null) {
            SyncDto syncDto = gson.fromJson(result, SyncDto.class);
@@ -274,6 +345,14 @@ public class SyncService {
                 oList = syncDto.getoList();
                 orgsService.batchInsert(oList);
                 LogUtil.i("oList", String.valueOf(oList.size()));
+            }
+            if(syncDto.getuList() != null && syncDto.getuList().size() > 0) {
+                uList = syncDto.getuList();
+                //Todo 添加人员批量插入接口
+            }
+            if (syncDto.getrList() != null && syncDto.getrList().size() > 0) {
+                rList = syncDto.getrList();
+                deviceService.batchInsertRele(rList);
             }
         }
